Actually write tests. Rename loaddata (cuz it breaks Django)
