+++ councilmatic_core/fixtures/event_document.json
@@ -1,0 +1,1 @@
+[{"model": "councilmatic_core.eventdocument", "pk": 1, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/5307044C-2B67-4636-BD0C-523290EAE243/Ageneda,%2012-7-2016.pdf", "full_text": null, "event": "ocd-event/8ea4ec9c-783e-44b0-b612-74fdfe69e957"}}, {"model": "councilmatic_core.eventdocument", "pk": 2, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/5307044C-2B67-4636-BD0C-523290EAE243/NotMeet%20WedDec%207,%202016%20T_20161102132912.pdf", "full_text": null, "event": "ocd-event/8ea4ec9c-783e-44b0-b612-74fdfe69e957"}}, {"model": "councilmatic_core.eventdocument", "pk": 3, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/C3DAB7E2-866E-4DCF-9EB7-982B60D1C608/Agenda,%2012-12-2016.pdf", "full_text": null, "event": "ocd-event/503a6a9d-18b6-49e9-ae61-a95175b67c5a"}}, {"model": "councilmatic_core.eventdocument", "pk": 4, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/C3DAB7E2-866E-4DCF-9EB7-982B60D1C608/notfin_20161117131927.pdf", "full_text": null, "event": "ocd-event/503a6a9d-18b6-49e9-ae61-a95175b67c5a"}}, {"model": "councilmatic_core.eventdocument", "pk": 5, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/2B108BEA-1F92-4EAF-BDBE-CC0BD3C4FF85/Agenda,%2012-7-2016.pdf", "full_text": null, "event": "ocd-event/b0f7e500-7a90-49e5-924d-27b9e2c64a14"}}, {"model": "councilmatic_core.eventdocument", "pk": 6, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/2B108BEA-1F92-4EAF-BDBE-CC0BD3C4FF85/Notice,%2012-7-2016.pdf", "full_text": null, "event": "ocd-event/b0f7e500-7a90-49e5-924d-27b9e2c64a14"}}, {"model": "councilmatic_core.eventdocument", "pk": 7, "fields": {"note": "Summary", "url": "http://media.legistar.com/chic/meetings/2B108BEA-1F92-4EAF-BDBE-CC0BD3C4FF85/Summary,%2012-7-2016.pdf", "full_text": null, "event": "ocd-event/b0f7e500-7a90-49e5-924d-27b9e2c64a14"}}, {"model": "councilmatic_core.eventdocument", "pk": 8, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/4D6830B7-AF41-450B-95C6-5784E8E49B78/Budget_20161208161311.pdf", "full_text": null, "event": "ocd-event/75d28740-4c80-42c2-ae15-31c69d4a9e85"}}, {"model": "councilmatic_core.eventdocument", "pk": 9, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/4D6830B7-AF41-450B-95C6-5784E8E49B78/Notice,%2012-13-2016.pdf", "full_text": null, "event": "ocd-event/75d28740-4c80-42c2-ae15-31c69d4a9e85"}}, {"model": "councilmatic_core.eventdocument", "pk": 10, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/1467A718-CBBD-4B68-9C6F-BDA4263B0E98/Zoning%2012-6%20cancel_20161202115733.pdf", "full_text": null, "event": "ocd-event/1c15aa99-dad9-4c2f-bcba-0eccc00d2032"}}, {"model": "councilmatic_core.eventdocument", "pk": 11, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/BB5971E3-8C5F-485B-95A0-02BF15071CA9/Agenda,%2012-5-2016.pdf", "full_text": null, "event": "ocd-event/ef26f8db-4dc4-4818-bb33-544327445216"}}, {"model": "councilmatic_core.eventdocument", "pk": 12, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/BB5971E3-8C5F-485B-95A0-02BF15071CA9/Notice,%2012-5-2016.pdf", "full_text": null, "event": "ocd-event/ef26f8db-4dc4-4818-bb33-544327445216"}}, {"model": "councilmatic_core.eventdocument", "pk": 13, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/CD8C7C7D-486E-4DC2-9467-C13C7A1C948D/Amended%20Agenda,%2011-29-2016.pdf", "full_text": null, "event": "ocd-event/0fbde7ab-f84a-48aa-9cc0-aac08266dc83"}}, {"model": "councilmatic_core.eventdocument", "pk": 14, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/CD8C7C7D-486E-4DC2-9467-C13C7A1C948D/Notice,%2011-29-2016.pdf", "full_text": null, "event": "ocd-event/0fbde7ab-f84a-48aa-9cc0-aac08266dc83"}}, {"model": "councilmatic_core.eventdocument", "pk": 15, "fields": {"note": "Summary", "url": "http://media.legistar.com/chic/meetings/CD8C7C7D-486E-4DC2-9467-C13C7A1C948D/Summary,%2011-29-2016.pdf", "full_text": null, "event": "ocd-event/0fbde7ab-f84a-48aa-9cc0-aac08266dc83"}}, {"model": "councilmatic_core.eventdocument", "pk": 16, "fields": {"note": "Cancellation", "url": "http://media.legistar.com/chic/meetings/6C9DEFB8-8CBF-40B5-847F-8A8DF81245B9/Zoning%20CANCEL%2011-2-2015_20151028142525(1).pdf", "full_text": null, "event": "ocd-event/3ba311a7-be8f-4b70-90df-15921b267251"}}, {"model": "councilmatic_core.eventdocument", "pk": 17, "fields": {"note": "Amended Agenda", "url": "http://media.legistar.com/chic/meetings/B0648725-C94D-4029-90F5-62543B51B25D/Notice%20and%20Amended%20Agenda,%2012-07-2016.pdf", "full_text": null, "event": "ocd-event/d077f7fb-fecb-4ab5-9ac3-8b3efd1505d1"}}, {"model": "councilmatic_core.eventdocument", "pk": 18, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/4C61B3AC-94B8-491E-B361-9E23C371074D/Agenda,%202014-2018%20Chicago%205%20Yr%20Hsg%20Plan%20Optimized.pdf", "full_text": null, "event": "ocd-event/e4f07825-2f73-40e8-8c8f-50aa8118774a"}}, {"model": "councilmatic_core.eventdocument", "pk": 19, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/4C61B3AC-94B8-491E-B361-9E23C371074D/NOT%20Meet%20HOUS%20RE%20Dec13.D_20161202081152.pdf", "full_text": null, "event": "ocd-event/e4f07825-2f73-40e8-8c8f-50aa8118774a"}}, {"model": "councilmatic_core.eventdocument", "pk": 20, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/D2FE4B0F-098E-4ED3-A54F-288BC441A480/Agenda,%2012-8-2016.pdf", "full_text": null, "event": "ocd-event/30fdd652-5765-4862-8351-5affe226218e"}}, {"model": "councilmatic_core.eventdocument", "pk": 21, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/D2FE4B0F-098E-4ED3-A54F-288BC441A480/128notped_20161102153448.pdf", "full_text": null, "event": "ocd-event/30fdd652-5765-4862-8351-5affe226218e"}}, {"model": "councilmatic_core.eventdocument", "pk": 22, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/E6DC2CF5-E081-4925-83D5-A57CB6E5BF65/Agenda%2011-16-2016.pdf", "full_text": null, "event": "ocd-event/8a34a02a-2b7c-4507-b970-faef53c19b69"}}, {"model": "councilmatic_core.eventdocument", "pk": 23, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/E6DC2CF5-E081-4925-83D5-A57CB6E5BF65/11-16-2016.pdf", "full_text": null, "event": "ocd-event/8a34a02a-2b7c-4507-b970-faef53c19b69"}}, {"model": "councilmatic_core.eventdocument", "pk": 24, "fields": {"note": "Transcript", "url": "https://chicago.legistar.com/Transcript.aspx?ID1=587", "full_text": null, "event": "ocd-event/8a34a02a-2b7c-4507-b970-faef53c19b69"}}, {"model": "councilmatic_core.eventdocument", "pk": 25, "fields": {"note": "Agenda", "url": "http://media.legistar.com/chic/meetings/E09E3AAF-5E3C-4CD8-A18B-F4FAE71C84D3/Agenda,%2012-12-2016.pdf", "full_text": null, "event": "ocd-event/a6287f14-ab96-460e-955c-2b31d150449a"}}, {"model": "councilmatic_core.eventdocument", "pk": 26, "fields": {"note": "Notice", "url": "http://media.legistar.com/chic/meetings/E09E3AAF-5E3C-4CD8-A18B-F4FAE71C84D3/Notice,%2012-12-2016.pdf", "full_text": null, "event": "ocd-event/a6287f14-ab96-460e-955c-2b31d150449a"}}]
