+++ councilmatic_core/fixtures/event_participant.json
@@ -1,0 +1,1 @@
+[{"model": "councilmatic_core.eventparticipant", "pk": 1, "fields": {"event": "ocd-event/8ea4ec9c-783e-44b0-b612-74fdfe69e957", "note": "participant", "entity_name": "Committee on Transportation and Public Way", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 2, "fields": {"event": "ocd-event/503a6a9d-18b6-49e9-ae61-a95175b67c5a", "note": "participant", "entity_name": "Committee on Finance", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 3, "fields": {"event": "ocd-event/b0f7e500-7a90-49e5-924d-27b9e2c64a14", "note": "participant", "entity_name": "Committee on License and Consumer Protection", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 4, "fields": {"event": "ocd-event/75d28740-4c80-42c2-ae15-31c69d4a9e85", "note": "participant", "entity_name": "Committee on Budget and Government Operations", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 5, "fields": {"event": "ocd-event/1c15aa99-dad9-4c2f-bcba-0eccc00d2032", "note": "participant", "entity_name": "Committee on Zoning, Landmarks and Building Standards", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 6, "fields": {"event": "ocd-event/ef26f8db-4dc4-4818-bb33-544327445216", "note": "participant", "entity_name": "Committee on Human Relations", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 7, "fields": {"event": "ocd-event/0fbde7ab-f84a-48aa-9cc0-aac08266dc83", "note": "participant", "entity_name": "Committee on Housing and Real Estate", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 8, "fields": {"event": "ocd-event/3ba311a7-be8f-4b70-90df-15921b267251", "note": "participant", "entity_name": "Committee on Zoning, Landmarks and Building Standards", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 9, "fields": {"event": "ocd-event/d077f7fb-fecb-4ab5-9ac3-8b3efd1505d1", "note": "participant", "entity_name": "Committee on Economic, Capital and Technology Development", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 10, "fields": {"event": "ocd-event/e4f07825-2f73-40e8-8c8f-50aa8118774a", "note": "participant", "entity_name": "Committee on Housing and Real Estate", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 11, "fields": {"event": "ocd-event/30fdd652-5765-4862-8351-5affe226218e", "note": "participant", "entity_name": "Committee on Pedestrian and Traffic Safety", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 12, "fields": {"event": "ocd-event/8a34a02a-2b7c-4507-b970-faef53c19b69", "note": "participant", "entity_name": "Chicago City Council", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}, {"model": "councilmatic_core.eventparticipant", "pk": 13, "fields": {"event": "ocd-event/a6287f14-ab96-460e-955c-2b31d150449a", "note": "participant", "entity_name": "Committee on Zoning, Landmarks and Building Standards", "entity_type": "organization", "updated_at": "2016-12-12T17:33:04.839Z"}}]
