+++ councilmatic_core/fixtures/legislative_session.json
@@ -1,0 +1,1 @@
+[{"model": "councilmatic_core.legislativesession", "pk": "2007", "fields": {"jurisdiction_ocd_id": "ocd-jurisdiction/country:us/state:il/place:chicago/government", "name": "2007 Legislative Session", "updated_at": "2016-12-12T17:30:46.010Z"}}, {"model": "councilmatic_core.legislativesession", "pk": "2011", "fields": {"jurisdiction_ocd_id": "ocd-jurisdiction/country:us/state:il/place:chicago/government", "name": "2011 Legislative Session", "updated_at": "2016-12-12T17:30:46.013Z"}}, {"model": "councilmatic_core.legislativesession", "pk": "2015", "fields": {"jurisdiction_ocd_id": "ocd-jurisdiction/country:us/state:il/place:chicago/government", "name": "2015 Legislative Session", "updated_at": "2016-12-12T17:30:46.017Z"}}]
