+++ councilmatic_core/tests.py
@@ -1,5 +1,11 @@
-from django.test import TestCase
+from django.test import TestCase, Client
+from django.db import connection
 
 
 class RouteTest(TestCase):
-    pass
+    fixtures = ['bill.json']
+
+    def test_index(self):
+        client = Client()
+        response = client.get('/')
+        assert response.status_code == 200
